fix compile errors

+++ src/testcurl/https/tls_alert_test.c
@@ -59,7 +59,7 @@ test_alert_close_notify (MHD_gtls_session_t session)
   sa.sin_port = htons (DEAMON_TEST_PORT);
   inet_pton (AF_INET, "127.0.0.1", &sa.sin_addr);
 
-  MHD__gnutls_transport_set_ptr (session, (MHD_gnutls_transport_ptr_t) sd);
+  MHD__gnutls_transport_set_ptr (session, (MHD_gnutls_transport_ptr_t) (long) sd);
 
   ret = connect (sd, &sa, sizeof (struct sockaddr_in));
 
@@ -114,7 +114,7 @@ test_alert_unexpected_message (MHD_gtls_session_t session)
   inet_pton (AF_INET, "127.0.0.1", &sa.sin_addr);
 
   MHD__gnutls_transport_set_ptr (session,
-                                 (MHD_gnutls_transport_ptr_t) ((void *) sd));
+                                 (MHD_gnutls_transport_ptr_t) ((void *) (long) sd));
 
   ret = connect (sd, &sa, sizeof (struct sockaddr_in));
